Extract coercion logic to _coercions.py with Protocol-based types
Move all heterogeneous-type coercion and detection helpers from
_core.py into a dedicated _coercions.py module. Use Protocol types
(_CheckFn, _CoerceFn) for _CoercionStep fields so existing functions
can be passed directly without lambda wrappers. Promote check
functions from nested closures to top-level module functions.

This reduces _core.py from ~2030 to ~1350 lines and makes the
coercion subsystem independently navigable.

Closes #974
